Modify and add sample program for PiperWaveProvider
